Second optimization (#2)
* Second round of optimizations

This reorders the loops that iterate through the parmeters used to
calculate S, and moves more of the calculation outside of the
inner loop.

* Removed the "MAX(a,b)" from the init_fpt function

Currently the MAX(a,b) is being done outside the init_fpt
function so this really did nothing.  Not a hugh change
probably won't be able to measure it - however it decrease
things by 1 comparision for each value of fptilde above fptildemin

* Reorganized

Corrected the Optimization mispelling
split the Dockerfile to a base image that can be shared between
other dockerfile and one that is specific for p3.

* A base docker file for c and fortran programs

Since this will be shared between 2 question, have added the basic
development tools in the base image that can be used to derive
the other dockerfiles from.

* Some Reorganization

Move git clone of the test001 project to the c-fortran-dev
dockerfile/container.  This way the directory structure is
created that the other dockerfiles (with c and fortran code)
 can reference

* Inital form of this dockerfile

Need to commit things in order to get the directory
to be named correctly

---------

Co-authored-by: rob-baron <robertbartlettbaron@gmail.com>